Move DataSourceFeatureRequirementBuildItem producers to HibernateOrmDevServicesProcessor processor
